Update Watch Code API
